Move schemas to /datasources
