+++ src/chains/chains.controller.spec.ts
@@ -3,7 +3,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import * as request from 'supertest';
 import chainFactory from '../datasources/config-api/entities/__tests__/chain.factory';
 import { ChainsModule } from './chains.module';
-import { Backbone, Chain } from './entities';
+import { Backbone } from './entities';
 import backboneFactory from './entities/__tests__/backbone.factory';
 import {
   mockNetworkService,
@@ -18,6 +18,7 @@ import {
   TestCacheModule,
 } from '../common/cache/__tests__/test.cache.module';
 import { Page } from '../common/entities/page.entity';
+import { Chain } from '../datasources/config-api/entities/chain.entity';
 
 describe('Chains Controller (Unit)', () => {
   let app: INestApplication;
