+++ src/chains/entities/index.ts
@@ -1,3 +1,1 @@
 export * from './backbone.entity';
-export * from './chain.entity';
-export * from './native.currency.entity';
