+++ src/datasources/config-api/entities/__tests__/chain.factory.ts
@@ -1,5 +1,6 @@
 import { faker } from '@faker-js/faker';
-import { Chain, NativeCurrency } from '../../../../chains/entities';
+import { Chain } from '../chain.entity';
+import { NativeCurrency } from '../native.currency.entity';
 import nativeCurrencyFactory from './native.currency.factory';
 
 export default function (
