+++ src/datasources/transaction-api/transaction-api.service.ts
@@ -4,8 +4,8 @@ import { CacheFirstDataSource } from '../cache/cache.first.data.source';
 import { ValidationErrorFactory } from '../errors/validation-error-factory';
 import { DefinedError, ValidateFunction } from 'ajv';
 import { JsonSchemaService } from '../../common/schemas/json-schema.service';
-import { balanceSchema } from '../../balances/entities/schemas/balance.schema';
-import { backboneSchema } from '../../chains/entities/schemas/backbone.schema';
+import { balanceSchema } from './entities/schemas/balance.schema';
+import { backboneSchema } from './entities/schemas/backbone.schema';
 import { tokenInfoSchema } from '../../common/schemas/token-info.schema';
 
 export class TransactionApi {
